LG-6872: Creation USPS enrollment when IPP user re-enters their password (#6615)
* Prepping to create USPS enrollment

* Begin calling usps

Broken -- tries to call the real USPS which we don't have configured

* Add mock USPS proofer and create enrollment

* Add mock service

* Move usps integration to password controller

* Use a more unique unique_id

* Add usps in person proofing directory

* Move structs into usps proofing directory

* Fix broken specs

* Move and simplify

* Fix lint failures

* Use a different unique ID

* Reorganize enrollment creation logic

* Remove unused changes

* Flesh out specs

* Use new namespace and name

* changelog: Upcoming features, In-person proofing, Begin creating enrollments in USPS api when user completes flow

* Reorganize specs and code

* Add another spec, remove some todos

* Remove unnecessary code and comments

* Slight refactor
