+++ app/services/usps_in_person_proofing/applicant.rb
@@ -1,0 +1,6 @@
+module UspsInPersonProofing
+  Applicant = Struct.new(
+    :unique_id, :first_name, :last_name, :address, :city, :state, :zip_code,
+    :email, keyword_init: true
+  )
+end
