+++ app/services/usps_in_person_proofing/post_office.rb
@@ -1,0 +1,5 @@
+module UspsInPersonProofing
+  PostOffice = Struct.new(
+    :distance, :address, :city, :phone, :name, :zip_code, :state, keyword_init: true
+  )
+end
